refactor!: remove tunnel, add signal handlers
