+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "quincy"
-version = "0.7.1"
+version = "0.8.0"
 authors = ["Jakub Kubík <jakub.kubik.it@protonmail.com>"]
 license = "MIT"
 description = "QUIC-based VPN"
@@ -51,6 +51,7 @@ tokio = { version = "^1.25", features = [
     "macros",
     "sync",
     "io-util",
+    "signal",
 ] }
 dashmap = "^5.4"
 futures = "^0.3.17"
