+++ examples/server.toml
@@ -1,6 +1,4 @@
-# Section representing tunnel configuration
-[tunnels.tun0]
-# Name of the tunnel (currently not used as the name of the interface)
+# Name of the server instance (currently not used as the name of the interface)
 name = "tun0"
 # Path to the certificate used for TLS
 certificate_file = "examples/cert/server_cert.pem"
