+++ src/utils.rs
@@ -1,5 +1,6 @@
 pub mod certificates;
 pub mod cli;
+pub mod signal_handler;
 pub mod socket;
 pub mod tasks;
 pub mod tracing;
